all features for comments
Each comment can be edited by the owner of the comment or an admin. Each comment can be deleted by the owner of the comment, an admin or the owner of the blog which the comment belongs to. Comments can be sorted by popularity (amount of likes, default) or by newest comments. The owner of the comment can not like the same comment multiple times. The owner can like his own comments.